MAINT: Refactor IN110 and ZR89 classes to use GammaPositron class
Created the GammaPositron and GammaPositronDecay classes to make it so that only one
NuclearDecay is listed for each Decay of the ZR89 and IN110 isotopes.  In the future, this
probably means that a particular type of NuclearDecay will be associated with a particular
isotope and the distinction between NuclearDecay and Isotope can be dropped.