ENH: project the number of interactions so that we generally don't overrun the soft max
Testing between calculating the projection, and dealing with the reallocation of the vector, this
method was significantly faster.  Perhaps there could be an intelligent way to calculate this, in
a way that we don't do the math for every decay, but that's probably marginal relative to the
ray tracing anyway.