ENH: Change the daq model so that it does not copy
The initial implementation of the daq model, for simplicity, just copied the
entire vector for each process step.  Since the Interation class is quite heavy
this started to add a lot of overhead.  Clean this up, so that we keep a
persistent, single buffer, than we only delete once.  This still causes a copy
on the remaining elements, but this is assumed to be quite small.  This also now
puts the hits through the daq model's initial sorting stage so that hits are
time ordered in their output.  This is important, as it is easy to forget in
gray-daq to sort the input.  This work took several stages:

BUG: Set dropped flag appropriately in interaction types
BUG: fix up some process implementations
BUG: fix merge process
BUG: mostly fix coinc sorting
BUG: should have been in last commit
BUG: fixup coincidence sort
MAINT: Document the coincidence sorting a bit more
MAINT: remove now unused functions
MAINT: Break gray, but gray-daq works
BUG: return proper iterator for sorting.
BUG: don't overwrite events in Input
MAINT: Keep stats in filtering
BUG: fix up return value on dead time process
BUG: fix iterator invalidation issue in Input
MAINT: cleanup processes
MAINT: remove ProcessStream and ParallelOutputStream
MAINT: add back in stats
MAINT: refactor processing logic into InteractionStream
MAINT: refactor singles and coinc begin and end to always be correct for the output
MAINT: make gray usable again.
MAINT: use insertion sort instead of std::sort