MAINT: move vector source to begin standalone from scene
VectorSource required all sorts of weird hooks into the scene and kdtree.  Move
it to be independent of the reset of the tree by moving IntersectKdTree's
functionality into SceneDescription and creating an independent SceneDescription
for every vector source.  This removes vector sources from the geometry of the
scene entirely, which is intended, and has been documented in the command
descriptions.  This allows VectorSources to be negative as well.